+++ lib/lightning_web/live/channel_live/index.ex
@@ -28,7 +28,7 @@ defmodule LightningWeb.ChannelLive.Index do
         <LayoutComponents.header current_user={@current_user}>
           <:breadcrumbs>
             <LayoutComponents.breadcrumbs>
-              <LayoutComponents.breadcrumb_project_picker label={@project.name} />
+              <LayoutComponents.breadcrumb_project_picker project={@project} />
               <LayoutComponents.breadcrumb>
                 <:label>{@page_title}</:label>
               </LayoutComponents.breadcrumb>