+++ lib/lightning_web/live/workflow_live/index.ex
@@ -39,7 +39,7 @@ defmodule LightningWeb.WorkflowLive.Index do
         <LayoutComponents.header current_user={@current_user}>
           <:breadcrumbs>
             <LayoutComponents.breadcrumbs>
-              <LayoutComponents.breadcrumb_project_picker label={@project.name} />
+              <LayoutComponents.breadcrumb_project_picker project={@project} />
               <LayoutComponents.breadcrumb>
                 <:label>{@page_title}</:label>
               </LayoutComponents.breadcrumb>